feat: add validation, improve error handling

+++ src/user/user.controller.ts
@@ -6,19 +6,25 @@ import {
   Body,
   Put,
   Delete,
+  NotFoundException,
+  UseInterceptors,
+  ClassSerializerInterceptor,
+  BadRequestException,
 } from '@nestjs/common';
 import { CreateUserDto } from './dtos/create-user.dto';
 import { UserService } from './user.service';
 import { User } from './user.entity';
 import { UpdatePasswordDto } from './dtos/update-password.dto';
+import * as uuid from 'uuid';
 
+@UseInterceptors(ClassSerializerInterceptor)
 @Controller('user')
 export class UserController {
   constructor(private readonly userService: UserService) {}
 
   @Post()
   async createUser(@Body() body: CreateUserDto): Promise<User> {
-    const user = await this.userService.createUser(body.login, body.password);
+    const user = await this.userService.create(body.login, body.password);
     return user;
   }
 
@@ -29,8 +35,10 @@ export class UserController {
   }
 
   @Get('/:id')
-  async getUserById(@Param('id') id: string): Promise<User> {
+  async getUserById(@Param('id') id: string) {
+    if (!uuid.validate(id)) throw new BadRequestException(`Invalid id ${id}`);
     const user = await this.userService.findById(id);
+    if (!user) throw new NotFoundException(`User with id ${id} not found`);
     return user;
   }
 
@@ -39,12 +47,18 @@ export class UserController {
     @Param('id') id: string,
     @Body() body: UpdatePasswordDto,
   ): Promise<User> {
-    const user = await this.userService.updateUserPassword(id, body.password);
+    if (!uuid.validate(id)) throw new BadRequestException(`Invalid id ${id}`);
+    const user = await this.userService.updateUserPassword(
+      id,
+      body.oldPassword,
+      body.newPassword,
+    );
     return user;
   }
 
   @Delete('/:id')
   async deleteUser(@Param('id') id: string): Promise<string> {
+    if (!uuid.validate(id)) throw new BadRequestException(`Invalid id ${id}`);
     const deletedUserId = await this.userService.delete(id);
     return deletedUserId;
   }
