+++ src/user/user.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { User } from './user.entity';
 import { UserRepository } from './user.repository';
 
@@ -6,7 +10,7 @@ import { UserRepository } from './user.repository';
 export class UserService {
   constructor(private readonly userRepo: UserRepository) {}
 
-  createUser(login: string, password: string): Promise<User> {
+  create(login: string, password: string): Promise<User> {
     return this.userRepo.create(login, password);
   }
 
@@ -18,11 +22,29 @@ export class UserService {
     return this.userRepo.findById(id);
   }
 
-  async updateUserPassword(id: string, password: string): Promise<User> {
-    return this.userRepo.updatePassword(id, password);
+  async updateUserPassword(
+    id: string,
+    oldPassword: string,
+    newPassword: string,
+  ): Promise<User> {
+    const user = await this.findById(id);
+    if (!user) throw new NotFoundException('User not found');
+    if (user.password !== oldPassword)
+      throw new ForbiddenException('Old password is incorrect');
+
+    const updatedUser = new User({
+      ...user,
+      password: newPassword,
+      updated_at: new Date().toISOString(),
+      version: user.version + 1,
+    });
+
+    return this.userRepo.updatePassword(id, updatedUser);
   }
 
-  delete(id: string): Promise<string> {
+  async delete(id: string): Promise<string> {
+    const user = await this.findById(id);
+    if (!user) throw new NotFoundException('User not found');
     return this.userRepo.delete(id);
   }
 }
